Add support for check setup actions
